add formula and video insertion to bubble
