+++ assets/bubble/tooltip.styl
@@ -18,12 +18,12 @@ arrowWidth = 6px
     top: -1 * arrowWidth
 
   .ql-tooltip.ql-editing
-    .ql-link-editor
+    .ql-tooltip-editor
       display: block
     .ql-formats
       visibility: hidden
 
-  .ql-link-editor
+  .ql-tooltip-editor
     display: none
     input[type=text]
       background: transparent
@@ -32,7 +32,7 @@ arrowWidth = 6px
       font-size: 13px
       height: 100%
       outline: none
-      padding: 10px 50px 10px 20px
+      padding: 10px 20px
       position: absolute
       width: 100%
     a
